// ui/src/apis/proto/openpay/v1/ledger.ts
export enum LedgerAccountType {
  LEDGER_ACCOUNT_TYPE_UNSPECIFIED = "LEDGER_ACCOUNT_TYPE_UNSPECIFIED",
  LEDGER_ACCOUNT_TYPE_ASSET = "LEDGER_ACCOUNT_TYPE_ASSET",
  LEDGER_ACCOUNT_TYPE_LIABILITY = "LEDGER_ACCOUNT_TYPE_LIABILITY",
  LEDGER_ACCOUNT_TYPE_EQUITY = "LEDGER_ACCOUNT_TYPE_EQUITY",
  LEDGER_ACCOUNT_TYPE_INCOME = "LEDGER_ACCOUNT_TYPE_INCOME",
  LEDGER_ACCOUNT_TYPE_EXPENSE = "LEDGER_ACCOUNT_TYPE_EXPENSE",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export interface TrialBalanceLine {
  accountId: string;
  accountCode: string;
  type: LedgerAccountType;
  productId: string;
  currency: string;
  debits: string;
  credits: string;
  /** debits - credits. */
  rawBalance: string;
  /** The balance in the account's natural direction. */
  balance: string;
}